+++ android/src/main/java/com/theoplayer/integration/PlayerFacadeEventSubscription.kt
@@ -5,7 +5,7 @@ import com.theoplayer.android.api.event.EventListener
 
 @Suppress("FINITE_BOUNDS_VIOLATION_IN_JAVA")
 internal class PlayerFacadeEventSubscription<E : Event<*>> {
-  val listeners = linkedMapOf<EventListener<*>, (E) -> Unit>()
+  val listeners = linkedMapOf<EventListener<*>, (E, Boolean) -> Unit>()
   val interceptors = linkedSetOf<(E) -> Boolean>()
   var remove: () -> Unit = {}
 }